+++ app/Makefile
@@ -80,6 +80,7 @@ VARIANT_PARAM=COIN
 VARIANT_VALUES=$(COIN)
 
 INCLUDES_PATH += $(CURDIR)/src/common
+INCLUDES_PATH += $(CURDIR)/src/evm
 INCLUDES_PATH += $(BOLOS_SDK)/lib_cxng/src
 
 DEFINES += APP_BLINDSIGN_MODE_ENABLED
