+++ app/src/apdu_handler.c
@@ -25,10 +25,11 @@
 #include "addr.h"
 #include "app_main.h"
 #include "coin.h"
-#include "common/tx.h"
+#include "coin_evm.h"
+#include "tx.h"
 #include "crypto.h"
-#include "eth_addr.h"
-#include "eth_utils.h"
+#include "evm_addr.h"
+#include "evm_utils.h"
 #include "tx.h"
 #include "view.h"
 #include "view_internal.h"
