+++ app/src/coin.h
@@ -20,7 +20,6 @@ extern "C" {
 #endif
 
 #define CLA 0x06
-#define CLA_ETH 0xE0
 
 #include <stddef.h>
 #include <stdint.h>
@@ -34,15 +33,8 @@ extern "C" {
 #define HDPATH_3_DEFAULT (0u)
 #define HDPATH_4_DEFAULT (0u)
 
-#define HDPATH_ETH_0_DEFAULT (0x80000000u | 0x2cu)
-#define HDPATH_ETH_1_DEFAULT (0x80000000u | 0x3cu)
-
 #define HDPATH_0_TESTNET (0x80000000u | 0x2cu)
 #define HDPATH_1_TESTNET (0x80000000u | 0x1u)
-
-#define SECP256K1_SK_LEN 64u
-#define SECP256K1_PK_LEN 65u
-#define ETH_ADDR_LEN 20u
 
 typedef enum {
     addr_secp256k1 = 0,
@@ -50,26 +42,15 @@ typedef enum {
 
 #define VIEW_ADDRESS_OFFSET_SECP256K1 \
     (SECP256K1_PK_LEN + ADDRESS_PROTOCOL_SECP256K1_PAYLOAD_LEN + ADDRESS_PROTOCOL_LEN + 2)
-// omit the pubkey + 1-byte pubkey len + 1-byte address len
-#define VIEW_ADDRESS_OFFSET_ETH (SECP256K1_PK_LEN + 1 + 1)
 
 #define COIN_AMOUNT_DECIMAL_PLACES 18
 
 #define COIN_SUPPORTED_TX_VERSION 0
 
-// transaction is sent as a blob of rlp encoded bytes,
-#define P1_ETH_FIRST 0x00
-#define P1_ETH_MORE 0x80
-// eth address chain_code allowed valuec
-#define P2_NO_CHAINCODE 0x00
-#define P2_CHAINCODE 0x01
-
 #define INS_GET_VERSION 0x00
 #define INS_GET_ADDR_SECP256K1 0x01
 #define INS_SIGN_SECP256K1 0x02
-#define INS_SIGN_ETH 0x04
 #define INS_SIGN_RAW_BYTES 0x07
-#define INS_GET_ADDR_ETH 0x02
 
 #define MENU_MAIN_APP_LINE1 "Filecoin"
 #define MENU_MAIN_APP_LINE2 "Ready"
