+++ app/src/evm/evm_addr.c
@@ -17,7 +17,7 @@
 #include <stdio.h>
 
 #include "app_mode.h"
-#include "coin.h"
+#include "coin_evm.h"
 #include "crypto.h"
 #include "zxerror.h"
 #include "zxformat.h"
