+++ app/src/evm/parser_impl_evm.c
@@ -14,16 +14,16 @@
  *  limitations under the License.
  ********************************************************************************/
 
-#include "parser_impl_eth.h"
+#include "parser_impl_evm.h"
 
 #include <stdio.h>
 #include <zxmacros.h>
 
 #include "app_mode.h"
-#include "common/parser_common.h"
+#include "parser_common.h"
 #include "crypto.h"
-#include "eth_erc20.h"
-#include "eth_utils.h"
+#include "evm_erc20.h"
+#include "evm_utils.h"
 #include "parser_txdef.h"
 #include "rlp.h"
 #include "uint256.h"
