+++ app/src/fil_utils.h
@@ -23,7 +23,7 @@ extern "C" {
 
 #include "bignum.h"
 #include "cbor.h"
-#include "common/parser_common.h"
+#include "parser_common.h"
 #include "crypto.h"
 #include "parser_txdef.h"
 #include "zxformat.h"
