+++ app/src/parser.c
@@ -22,10 +22,10 @@
 #include "app_mode.h"
 #include "bignum.h"
 #include "coin.h"
-#include "common/parser_common.h"
+#include "parser_common.h"
 #include "fil_utils.h"
 #include "parser_impl.h"
-#include "parser_impl_eth.h"
+#include "parser_impl_evm.h"
 #include "parser_invoke_evm.h"
 #include "parser_raw_bytes.h"
 #include "parser_txdef.h"
